MYC-1304: Change Root Route to Initial Chat Interface
