+++ hooks/usePrompts.tsx
@@ -12,14 +12,14 @@ const usePrompts = () => {
   const [currentQuestion, setCurrentQuestion] = useState<Message | null>(null);
   const pathname = usePathname();
   const { funnelRawReportContent } = useFunnelReportProvider();
-  const isNewChat = pathname.includes("/new");
+  const isNewChat = pathname.includes("/new") || pathname === "/";
 
   useEffect(() => {
     if (isLoading) return;
     if (selectedArtist && isNewChat) {
       setPrompts([
-        `Who are ${selectedArtist?.name || ""}’s most engaged fans?`,
-        `Analyze ${selectedArtist?.name || ""}’s TikTok posts from this week.`,
+        `Who are ${selectedArtist?.name || ""}'s most engaged fans?`,
+        `Analyze ${selectedArtist?.name || ""}'s TikTok posts from this week.`,
       ]);
       return;
     }
